requirements.txt
streamlit
requests
pandas
python-dotenv
beautifulsoup4
spacy>=3.7.0,<3.8.0

# Data analysis and visualization
numpy
scikit-learn
plotly
networkx
pyvis
python-louvain

# Optional but recommended
matplotlib
seaborn

# Note: After installing, download the spaCy model:
# python -m spacy download xx_ent_wiki_sm
